VirtualGox Implimented
Basic implimentation of VirtualGox.
The virtual api interface. It's an extension of MtGoxClient, so you can sub it in anywhere that class can be used.

This allows you to easily sub out for testing purposes.
It overrides the getprivatedata and addorder functionality. All other functionality talks to the live mtgox api.
the getprivatedata call talks to live mtgox, but then the wallets are overridden with a local only "virtual" wallet, which is pre-loaded with 100 of each currency (BTC, USD, CAD).
addorder is 100% virtual. Right now it's simple. It checks if you have enough in the appropriate virtual balance, and if so it executes the order in full (it does not verify if market depth would have allowed the execution of the order, this is planned for the future).